requirements.txt
asyncserial
pytest
numpy
sipyco @ git+https://github.com/m-labs/sipyco.git@master#egg=sipyco
